refactor(home): move homepage into a dedicated module
