+++ vite.config.ts
@@ -15,7 +15,6 @@ export default defineConfig({
       '@components': fileURLToPath(new URL('./src/components', import.meta.url)),
       '@lib': fileURLToPath(new URL('./src/lib', import.meta.url)),
       '@modules': fileURLToPath(new URL('./src/modules', import.meta.url)),
-      '@pages': fileURLToPath(new URL('./src/pages', import.meta.url)),
       '@shared': fileURLToPath(new URL('./src/shared', import.meta.url)),
       '@test': fileURLToPath(new URL('./src/test', import.meta.url)),
       '@theme': fileURLToPath(new URL('./src/theme', import.meta.url)),
